Simplifying expvar integration (#696)
* Spliting the work for expvar in seperate files

Had rerwritten some parts to make it clear of its intent

* Trying to rationalise things further

* Moving some code around to better suit creation / usage
