+++ internal/monitors/expvar/expvar.go
@@ -4,19 +4,18 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"io/ioutil"
+	"net/http"
+	"net/url"
+	"strconv"
+	"strings"
+	"time"
+
 	"github.com/signalfx/golib/datapoint"
-	"github.com/signalfx/signalfx-agent/internal/core/config"
 	"github.com/signalfx/signalfx-agent/internal/monitors"
 	"github.com/signalfx/signalfx-agent/internal/monitors/types"
 	"github.com/signalfx/signalfx-agent/internal/utils"
 	log "github.com/sirupsen/logrus"
-	"io/ioutil"
-	"net/http"
-	"net/url"
-	"regexp"
-	"strconv"
-	"strings"
-	"time"
 )
 
 const monitorType = "expvar"
@@ -37,98 +36,21 @@ const (
 
 var logger = log.WithFields(log.Fields{"monitorType": monitorType})
 
-// Config for monitor configuration
-type Config struct {
-	config.MonitorConfig `yaml:",inline" acceptsEndpoints:"true"`
-	// Host of the expvar endpoint
-	Host string `yaml:"host" validate:"required"`
-	// Port of the expvar endpoint
-	Port uint16 `yaml:"port" validate:"required"`
-	// If true, the agent will connect to the host using HTTPS instead of plain HTTP.
-	UseHTTPS bool `yaml:"useHTTPS"`
-	// If useHTTPS is true and this option is also true, the host's TLS cert will not be verified.
-	SkipVerify bool `yaml:"skipVerify"`
-	// Path to the expvar endpoint, usually `/debug/vars` (the default).
-	Path string `yaml:"path" default:"/debug/vars"`
-	// If true, sends metrics memstats.alloc, memstats.by_size.size, memstats.by_size.mallocs and memstats.by_size.frees
-	EnhancedMetrics bool `yaml:"enhancedMetrics"`
-	// Metrics configurations
-	MetricConfigs []*MetricConfig `yaml:"metrics"`
-}
-
-// MetricConfig for metric configuration
-type MetricConfig struct {
-	// Metric name
-	Name string `yaml:"name"`
-	// JSON path of the metric value
-	JSONPath string `yaml:"JSONPath" validate:"required"`
-	// SignalFx metric type. Possible values are "gauge" or "cumulative"
-	Type string `yaml:"type" validate:"required"`
-	// Metric dimensions
-	DimensionConfigs []*DimensionConfig `yaml:"dimensions"`
-}
-
-// DimensionConfig for metric dimension configuration
-type DimensionConfig struct {
-	// Dimension name
-	Name string `yaml:"name"`
-	// JSON path of the dimension value
-	JSONPath string `yaml:"JSONPath"`
-	// Dimension value
-	Value string `yaml:"value"`
-}
-
-// Validate validates configuration
-func (conf *Config) Validate() error {
-	if conf.MetricConfigs != nil {
-		for _, mConf := range conf.MetricConfigs {
-			// Validating metric type configuration
-			metricTypeString := strings.TrimSpace(strings.ToLower(mConf.Type))
-			if metricTypeString != gauge && metricTypeString != cumulative {
-				return fmt.Errorf("unsupported metric type %s. Supported metric types are: %s, %s", mConf.Type, gauge, cumulative)
-			}
-			// Validating dimension configuration
-			for _, dConf := range mConf.DimensionConfigs {
-				switch {
-				case dConf == nil:
-					continue
-				case dConf.Name == "" || strings.ReplaceAll(dConf.Name, " ", "") != dConf.Name:
-					return fmt.Errorf("dimension name cannot be blank or have whitespaces")
-				case dConf.JSONPath != "" && dConf.Value != "":
-					return fmt.Errorf("dimension path %s and dimension value %s cannot be configure simultaneously", dConf.JSONPath, dConf.Value)
-				case dConf.JSONPath != "" && !strings.HasPrefix(mConf.JSONPath, dConf.JSONPath):
-					return fmt.Errorf("dimension path %s must be shorter than metric path %s and start from the same root", dConf.JSONPath, mConf.JSONPath)
-				}
-			}
+func init() {
+	monitors.Register(monitorType, func() interface{} {
+		return &Monitor{
+			metricTypes:         make(map[*MetricConfig]datapoint.MetricType),
+			metricPathsParts:    make(map[*MetricConfig][]string),
+			dimensionPathsParts: make(map[*DimensionConfig][]string),
+			allMetricConfigs:    []*MetricConfig{},
 		}
-	}
-	return nil
-}
-
-func (mConf *MetricConfig) name() string {
-	if strings.TrimSpace(mConf.Name) == "" {
-		return toSnakeCase(mConf.JSONPath)
-	}
-	return mConf.Name
-}
-
-func (mConf *MetricConfig) dimensions() map[string]string {
-	var dimensions map[string]string
-	if len(mConf.DimensionConfigs) > 0 {
-		dimensions = make(map[string]string, len(mConf.DimensionConfigs))
-		for _, dConf := range mConf.DimensionConfigs {
-			if strings.TrimSpace(dConf.Name) != "" && strings.TrimSpace(dConf.Value) != "" {
-				dimensions[dConf.Name] = dConf.Value
-			}
-		}
-	}
-	return dimensions
+	}, &Config{})
 }
 
 // Monitor for expvar metrics
 type Monitor struct {
 	Output              types.Output
-	cancel              func()
+	cancel              context.CancelFunc
 	ctx                 context.Context
 	client              *http.Client
 	url                 *url.URL
@@ -139,16 +61,8 @@ type Monitor struct {
 	allMetricConfigs    []*MetricConfig
 }
 
-func init() {
-	monitors.Register(monitorType, func() interface{} { return &Monitor{} }, &Config{})
-}
-
 // Configure monitor
 func (m *Monitor) Configure(conf *Config) error {
-	m.metricTypes = map[*MetricConfig]datapoint.MetricType{}
-	m.metricPathsParts = map[*MetricConfig][]string{}
-	m.dimensionPathsParts = map[*DimensionConfig][]string{}
-	m.allMetricConfigs = []*MetricConfig{}
 	m.addDefaultMetricConfigs(conf.EnhancedMetrics)
 	for _, mConf := range conf.MetricConfigs {
 		if m.metricTypes[mConf] = datapoint.Gauge; strings.TrimSpace(strings.ToLower(mConf.Type)) == cumulative {
@@ -164,7 +78,7 @@ func (m *Monitor) Configure(conf *Config) error {
 			}
 			return "http"
 		}(),
-		Host: conf.Host + ":" + fmt.Sprint(conf.Port),
+		Host: fmt.Sprintf("%s:%d", conf.Host, conf.Port),
 		Path: conf.Path,
 	}
 	m.runInterval = time.Duration(conf.IntervalSeconds) * time.Second
@@ -173,8 +87,8 @@ func (m *Monitor) Configure(conf *Config) error {
 		Transport: &http.Transport{
 			MaxIdleConnsPerHost: 10,
 		},
+		Timeout: 300 * time.Millisecond,
 	}
-	m.Output.AddExtraDimension("port", strconv.FormatInt(int64(conf.Port), 10))
 	utils.RunOnInterval(m.ctx, func() {
 		dpsMap, err := m.fetchMetrics(conf)
 		if err != nil {
@@ -183,7 +97,7 @@ func (m *Monitor) Configure(conf *Config) error {
 		}
 		mostRecentGCPauseIndex := getMostRecentGCPauseIndex(dpsMap)
 		now := time.Now()
-		for metricPath, dps := range *dpsMap {
+		for metricPath, dps := range dpsMap {
 			for _, dp := range dps {
 				if err := m.sendDatapoint(conf, dp, metricPath, mostRecentGCPauseIndex, &now); err != nil {
 					logger.Error(err)
@@ -224,7 +138,7 @@ func (m *Monitor) addDefaultMetricConfigs(enhancedMetrics bool) {
 	}
 }
 
-func (m *Monitor) fetchMetrics(conf *Config) (*map[string][]*datapoint.Datapoint, error) {
+func (m *Monitor) fetchMetrics(conf *Config) (map[string][]*datapoint.Datapoint, error) {
 	resp, err := m.client.Get(m.url.String())
 	if err != nil {
 		return nil, err
@@ -234,90 +148,90 @@ func (m *Monitor) fetchMetrics(conf *Config) (*map[string][]*datapoint.Datapoint
 	if err != nil {
 		return nil, err
 	}
-	metricsJSON := interface{}(nil)
+	metricsJSON := make(map[string]interface{})
 	err = json.Unmarshal(body, &metricsJSON)
 	if err != nil {
 		return nil, err
 	}
-	var dpsMap map[string][]*datapoint.Datapoint
-	var applicationName string
-	if aMap, ok := metricsJSON.(map[string]interface{}); ok {
-		if applicationName, err = getApplicationName(aMap); err != nil {
-			logger.Error(err)
-		}
-		dpsMap = make(map[string][]*datapoint.Datapoint)
-		for _, mConf := range m.allMetricConfigs {
-			dp := datapoint.Datapoint{Dimensions: map[string]string{}}
-			if applicationName != "" {
-				dp.Dimensions["application_name"] = applicationName
-			}
-			dpsMap[mConf.JSONPath] = make([]*datapoint.Datapoint, 0)
-			m.setDatapoints(aMap[m.metricPathsParts[mConf][0]], mConf, &dp, &dpsMap, 0)
-		}
+	applicationName, err := getApplicationName(metricsJSON)
+	if err != nil {
+		logger.Warn(err)
 	}
-	return &dpsMap, nil
+	dpsMap := make(map[string][]*datapoint.Datapoint)
+	for _, mConf := range m.allMetricConfigs {
+		dp := datapoint.Datapoint{Dimensions: map[string]string{}}
+		if applicationName != "" {
+			dp.Dimensions["application_name"] = applicationName
+		}
+		dpsMap[mConf.JSONPath] = make([]*datapoint.Datapoint, 0)
+		m.setDatapoints(metricsJSON[m.metricPathsParts[mConf][0]], mConf, &dp, dpsMap, 0)
+	}
+
+	return dpsMap, nil
 }
 
-// setDatapoints the dp argument should be a pointer to a zero value datapoint
-// setDatapoints traverses v recursively following metric path parts in mConf.keys[]
-// setDatapoints adds dimensions along the way and sets metric value in the end
-// setDatapoints clones datapoints and add array index dimension for array values in v
-func (m *Monitor) setDatapoints(v interface{}, mConf *MetricConfig, dp *datapoint.Datapoint, dpsMap *map[string][]*datapoint.Datapoint, metricPathIndex int) {
-	if metricPathIndex >= len(m.metricPathsParts[mConf]) {
-		logger.Errorf("failed to find metric value in path: %s", mConf.JSONPath)
+// setDatapoints the dp argument should be a pointer to a zero value datapoint and
+// traverses v recursively following metric path parts in mConf.keys[]
+// adds dimensions along the way and sets metric value in the end
+// clones datapoints and add array index dimension for array values in v
+func (m *Monitor) setDatapoints(v interface{}, mc *MetricConfig, dp *datapoint.Datapoint, dpsMap map[string][]*datapoint.Datapoint, metricPathIndex int) {
+	if metricPathIndex >= len(m.metricPathsParts[mc]) {
+		logger.Errorf("failed to find metric value in path: %s", mc.JSONPath)
 		return
 	}
-	if aMap, ok := v.(map[string]interface{}); ok {
-		for _, dConf := range mConf.DimensionConfigs {
+	switch v.(type) {
+	case nil:
+		logger.Errorf("failed to find value %s with JSON path %s", mc.name(), mc.JSONPath)
+		return
+	case map[string]interface{}:
+		for _, dConf := range mc.DimensionConfigs {
 			if len(m.dimensionPathsParts[dConf]) != 0 && len(m.dimensionPathsParts[dConf]) == metricPathIndex {
-				dp.Dimensions[dConf.Name] = m.metricPathsParts[mConf][metricPathIndex]
+				dp.Dimensions[dConf.Name] = m.metricPathsParts[mc][metricPathIndex]
 			}
 		}
-		m.setDatapoints(aMap[m.metricPathsParts[mConf][metricPathIndex+1]], mConf, dp, dpsMap, metricPathIndex+1)
-	} else if array, ok := v.([]interface{}); ok {
-		newDP := dp
-		for arrayIndex, arrayValue := range array {
-			if arrayIndex > 0 {
-				// At this point nothing is set for the do except possibly some dimensions
-				newDP = &datapoint.Datapoint{Dimensions: utils.CloneStringMap(dp.Dimensions)}
+		set := v.(map[string]interface{})
+		m.setDatapoints(set[m.metricPathsParts[mc][metricPathIndex+1]], mc, dp, dpsMap, metricPathIndex+1)
+	case []interface{}:
+		values := v.([]interface{})
+		clone := dp
+		for index, value := range values {
+			if index > 0 {
+				clone = &datapoint.Datapoint{Dimensions: utils.CloneStringMap(clone.Dimensions)}
 			}
-			makeNewIndexDimension := true
-			for _, dConf := range mConf.DimensionConfigs {
-				if len(m.dimensionPathsParts[dConf]) == metricPathIndex+1 {
-					newDP.Dimensions[dConf.Name] = strconv.Itoa(arrayIndex)
-					makeNewIndexDimension = false
+			createIndexDimension := true
+			for _, conf := range mc.DimensionConfigs {
+				if len(m.dimensionPathsParts[conf]) == metricPathIndex+1 {
+					clone.Dimensions[conf.Name] = fmt.Sprint(index)
+					createIndexDimension = false
 				}
 			}
-			if makeNewIndexDimension {
-				newDP.Dimensions[strings.Join(m.metricPathsParts[mConf][:(metricPathIndex+1)], ".")] = strconv.Itoa(arrayIndex)
+			if createIndexDimension {
+				clone.Dimensions[strings.Join(m.metricPathsParts[mc][:metricPathIndex+1], ".")] = fmt.Sprint(index)
 			}
-			m.setDatapoints(arrayValue, mConf, newDP, dpsMap, metricPathIndex)
+			m.setDatapoints(value, mc, clone, dpsMap, metricPathIndex)
 		}
-	} else {
-		if v == nil {
-			logger.Errorf("failed to find value for metric %s with JSON path %s", mConf.name(), mConf.JSONPath)
-			return
-		}
-		dp.Metric, dp.MetricType = mConf.name(), m.metricTypes[mConf]
-		for _, dConf := range mConf.DimensionConfigs {
+	default:
+		dp.Metric, dp.MetricType = mc.name(), m.metricTypes[mc]
+		for _, dConf := range mc.DimensionConfigs {
 			if strings.TrimSpace(dConf.Name) != "" && strings.TrimSpace(dConf.Value) != "" {
 				dp.Dimensions[dConf.Name] = dConf.Value
 			}
 		}
 		var err error
 		if dp.Value, err = datapoint.CastMetricValueWithBool(v); err == nil {
-			(*dpsMap)[mConf.JSONPath] = append((*dpsMap)[mConf.JSONPath], dp)
+			dpsMap[mc.JSONPath] = append(dpsMap[mc.JSONPath], dp)
 		} else {
-			logger.Errorf("failed to set value for metric %s with JSON path %s because of type conversion error due to %+v", mConf.name(), mConf.JSONPath, err)
+			logger.Debugf("failed to set value for metric %s with JSON path %s because of type conversion error due to %+v", mc.name(), mc.JSONPath, err)
+			logger.WithError(err).Error("Unable to set metric value")
 			return
 		}
 	}
 }
 
-func (m *Monitor) sendDatapoint(conf *Config, dp *datapoint.Datapoint, metricPath string, mostRecentGCPauseIndex int64, now *time.Time) (err error) {
+func (m *Monitor) sendDatapoint(conf *Config, dp *datapoint.Datapoint, metricPath string, mostRecentGCPauseIndex int64, now *time.Time) error {
 	if metricPath == memstatsPauseNsMetricPath || metricPath == memstatsPauseEndMetricPath {
-		var index int64
-		if index, err = strconv.ParseInt(dp.Dimensions[metricPath], 10, 0); err == nil && index == mostRecentGCPauseIndex {
+		index, err := strconv.ParseInt(dp.Dimensions[metricPath], 10, 0)
+		if err == nil && index == mostRecentGCPauseIndex {
 			dp.Metric = memstatsMostRecentGCPauseNsMetricName
 			if metricPath == memstatsPauseEndMetricPath {
 				dp.Metric = memstatsMostRecentGCPauseEndMetricName
@@ -326,7 +240,7 @@ func (m *Monitor) sendDatapoint(conf *Config, dp *datapoint.Datapoint, metricPat
 			delete(dp.Dimensions, metricPath)
 		} else {
 			if err != nil {
-				err = fmt.Errorf("failed to set metric most recent GC pause metric. %+v", err)
+				err = fmt.Errorf("failed to set metric GC pause. %+v", err)
 			}
 			return err
 		}
@@ -346,45 +260,4 @@ func (m *Monitor) Shutdown() {
 	if m.cancel != nil {
 		m.cancel()
 	}
-}
-
-// getMostRecentGCPauseIndex logic is derived from https://golang.org/pkg/runtime/ in the PauseNs section of the 'type MemStats' section
-func getMostRecentGCPauseIndex(dpsMap *map[string][]*datapoint.Datapoint) int64 {
-	dps := (*dpsMap)[memstatsNumGCMetricPath]
-	mostRecentGCPauseIndex := int64(-1)
-	if len(dps) > 0 && dps[0].Value != nil {
-		if numGC, err := strconv.ParseInt(dps[0].Value.String(), 10, 0); err == nil {
-			mostRecentGCPauseIndex = (numGC + 255) % 256
-		}
-	}
-	return mostRecentGCPauseIndex
-}
-
-var slashLastRegexp = regexp.MustCompile("[^\\/]*$")
-
-func getApplicationName(aMap map[string]interface{}) (string, error) {
-	if cmdline, ok := aMap["cmdline"].([]interface{}); ok && len(cmdline) > 0 {
-		if applicationName := strings.TrimSpace(slashLastRegexp.FindStringSubmatch(cmdline[0].(string))[0]); applicationName != "" {
-			return applicationName, nil
-		}
-	}
-	return "", fmt.Errorf("failed to get application name from the first array value of cmdline in map: %+v", aMap)
-}
-
-var camelRegexp = regexp.MustCompile("(^[^A-Z]*|[A-Z]*)([A-Z][^A-Z]+|$)")
-
-func toSnakeCase(s string) string {
-	snake := ""
-	for _, split := range strings.Split(s, ".") {
-		for _, submatches := range camelRegexp.FindAllStringSubmatch(split, -1) {
-			for _, submatch := range submatches[1:] {
-				submatch = strings.TrimSpace(submatch)
-				if submatch != "" {
-					snake += submatch + "_"
-				}
-			}
-		}
-		snake = strings.TrimSuffix(strings.TrimSuffix(snake, "."), "_") + "."
-	}
-	return strings.ToLower(strings.TrimSuffix(snake, "."))
 }
